Use %lu instead of %zu for format strings
The code normally uses %zu to print out size_t variables, but our
MinGW32 installation doesn't like that, so let's use %lu
(here, size_t == long unsigned int == unsigned long)

+++ bcinfo/MetadataExtractor.cpp
@@ -388,8 +388,8 @@ bool MetadataExtractor::populateForEachMetadata(
     }
   } else {
     if (mExportForEachSignatureCount != 1) {
-      ALOGE("mExportForEachSignatureCount = %zu, but should be 1",
-            mExportForEachSignatureCount);
+      ALOGE("mExportForEachSignatureCount = %lu, but should be 1",
+            (unsigned long) mExportForEachSignatureCount);
     }
     char *RootName = new char[5];
     strncpy(RootName, "root", 5);
